requirements.txt
numpy==1.11.0
networkx==1.11
-e git+https://github.com/tejado/pgoapi.git@81e786cabf027a1c8fbd1e9a07e1c11aa3d8ee8b#egg=pgoapi
geopy==1.11.0
protobuf==3.0.0b4
requests==2.10.0
s2sphere==0.2.4
gpsoauth==0.3.0
protobuf-to-dict==0.1.0
googlemaps==2.4.4
colorama==0.3.7
enum34==1.1.6
pyyaml==3.11
haversine==0.4.5
polyline==1.3.1
python-socketio==1.4.2
flask==0.11.1
socketIO_client==0.7.0
eventlet==0.19.0
universal-analytics-python==0.2.4
gpxpy==1.1.1
mock==2.0.0
timeout-decorator==0.3.2
python-dateutil==2.5.2
Flask==0.11.1
